Exclude grpc fault interceptor from production builds
